Add event and INFO level logging.
This introduces a few features:
* INFO level logging on common events like creating or fetching
  submissions and assessments.
* Parallel event logging for the same.
* Creates a local logs/ directory for development.

For the most part, we rely on XBlock's publish() mechanism to emit
event data. However, because scoring is done at the API layer in a way
that is not directly visible to the OpenAssesmentBlock, we're logging
that event separately with a hacky solution that dynamically loads an
emit function based on a configuration value. This is a stopgap
measure until we can start using the edX analytics API (which is still
in testing).

TIM-260, TIM-378