TIM-410: Make rubric prompt display optional
If users want rich prompts, they can do that with an HTML Components
above the ORA2 question, but doing this makes the prompt box look bad.
This adds a check so that the prompt box will not display if the
<prompt> XML tag is missing is empty or whitespace-only. Also extends
the serializer to serialize missing prompts distinctly from empty
prompts.

Includes tests for each of these conditions.