Instructors can edit default response feedback text
Previously, instructors were unable to edit the feedback prompt text ("I noticed that
this response...") displayed in the response box asking a student to provide feedback
on the peer responses they received.

To allow instructors to edit this, I've added an editable xml element (feedback_default_text).
The text within the element defaults to "I noticed that this response...," but an instructor
can change this text to whatever he or she wants by editing the text within the xml element.

Editing of this new element has also been added to the authoring gui.