Allow file upload to backends other than S3
We would like to support open assessment file uploads for open-edx
installs without access to S3, e.g: to the local fileystem. To that end,
we propose to abstract away some of the logic behind the file upload
backends so that other backends may be implemented.

In this proposed implementation, the uploaded files are stored in the
following folder structure:

        root/bucket/prefix/
                filename.jpg/
                        content
                        metadata.json

Note that a couple new settings need to be defined in order to use this
feature; a new urls.py file must also be included.