LaTeX not rendering properly in peer assessment
Latex was not working properly in peer assessment
as Mathjax was working with Latex on the element and
Mathjax completed its process before the peer response
that was not letting the element to render according to
Latex and Mathjax. Now rendering of Latex and Mathjax
will be done once the Mathjax will be loaded completely
and process the whole page.

TNL-1350