Fix up issues with peer scores
The crux of this fix is that scored_only behavior in various api/peer.py
methods is no longer needed, or correct. This value has been removed, so
that assessments may be gotten (and medians found) when the assessments
are not used in determining the final score (say, if a staff score exists).

Also includes updates to the "waiting for steps" string as prescribed by
@catong in https://github.com/edx/edx-ora2/pull/764, along with logic for
that method, and updates to tests and get_wating_details logic in light
of this simplification.