Allow openassessment xblock to clear state
All the changes here are in support of one goal - openassessmentblock
should be able to provide a clear_state method which will:
* soft-delete any models that refer to a submission_uuid
* passthrough to the submissions api to delete the submission as well

Doing this will allow staff the ability to reset student state on ORA problems,
which is otherwise impossible. The changes in this commit are an entry point
in openassessmentblock, helpers in workflow_ and submission_mixin, and changes to
the workflow and assessment models and apis to allow them to be soft deleted.

Migrations are needed as a part of this change, they too are included.