Override submitter requirements on staff override assessment
TNL-4696

When a staff override assessment is recorded, we now treat all the submitter's
requirements as being fulfilled. This allows ORA tooling to match the progress
page reported grade for the student by placing the AssessmentWorkflow into a
DONE state.

Acceptance tests have been updated, and are now also included in quality checks