resource_string provides a byte string.
Docs here: https://setuptools.readthedocs.io/en/latest/pkg_resources.html

The resource string provides bytes but we should convert to unicode as
soon as we get it for passing around purposes.

This data is eventually added as part of a json response so having a
byte string breaks that far downstream.