EDUCATOR-4810: Team submission generates submission records for all team members (#1334)
* Version bump to 2.6.6

* Move team submissions into a standalone function

* Expand team mixin to surface team ID

* Use anonymous user IDs for team file submisisons

* Get submission ID from DB if not stored locally

* Fix SharedFileUpload field access

* Wrap submission ID lookup errors to avoid first-time access UI bug

When first loading a team-enabled ORA, there is a StudentItem
lookup. This, however, is called for each section, producing
multiple calls on page load. This leads to a race condition where
Submissions will attempt to save multiple StudentItems which
causes the UI to show "unable to load" warnings. Internally,
errors are still thrown due to duplicate entries, but these are
not surfaced to the user and doesn't seem to actually cause any
issues outside of gross error logs.

* Get individual submission UUID while submitting