feat: ESG view team submissions (#1816)
* feat: require_submission_uuid for teams

When we validate submissions using @require_submission_uuid with
validate=True, we need to be sensitive to what type of submission (individual
vs team) we're validating.

* feat: update list_staff_workflows for teams

* Remove 400 error for using on team ORAs
* Make bulk lookup of submissions teams aware
* Fetch different context for serialization of teams submission list

* feat: update get_submission_info for teams

* Return based off of team submission ID for team assignments
* Behavior unchanged for individual assignments

* feat: update get_assessment_info for teams

* Return based off of team submission ID for team assignments
* Behavior unchanged for individual assignments

* feat: serialize team submission lists

* test: add tests for team list serialization

* test: update decorator tests

* test: add basic testing for team submission list

* chore: bump versions

* refactor: determine is_team_assignemnt only once

We do several tests on query counts, but the is_team_assignment in many
of our internal functions breaks those assertions. While this still
feels like we're testing/mocking too deeply, this allows us to keep
existing tests and cut out a few queries.