fix: upgrade pip-tools to fix bug in version 6.6.1
This commit upgrades the version of pip-tools used in this repository from 6.6.1 to 6.6.2.

In version 6.6.1 of pip-tools, there is a bug that is preventing pip-tools from working. This is breaking the Python requirements update GitHub action in this repository.

The error is "AttributeError: 'RequirementSet' object has no attribute 'add_requirement'". The error was reported here: https://github.com/jazzband/pip-tools/issues/1623. The fix to this bug was released in version 6.6.2 of pip-tools. See the pull request here: https://github.com/jazzband/pip-tools/pull/1624.