fix: don't check if submission_uuid from bulk_fetch_workflow matches
For team submissions, we request this function using the team submission uuid. We look up the TeamStaffWorkflow for this submission uuid, and then get the assessment associated with this workflow. The problem here is that that assessment is for one of the individual submissions, so the submission uuid on the assessment won't be the team submission uuid. This is a fix so that grading can happen, but we wil need a more robust solution, since the grade not howing up on the listing page is this exact same issue, but without this workaround.