test: Remove some test constraints that point to a merged PR.
It's not clear why these constraints point to this PR which was merged 4
years ago so remove them and see if we run into any issues when running
tests.

Leaving these here is causing an issue where make upgrade can't upgrade
anymore because newer versions of `tox` have a dependency on `pluggy` at
a higher version than what `pytest==4.5.0` is compatible with.  This is
what led me to look at these constraints in the first place.