fix: Update logger to output the uuid as a string.
In one of the tests the value of the ID is a mock.  In newer versions of
mock, the value won't get auto coerced to a number.  Also, it doesn't
make sense to coerce the UUID as a number so updated the logging to
output the UUID as a string in general.