feat: allow autosave to recover from save error
Previously, we wouldn't autosave if we had previously encountered an
error because the user could manually retry. On removing the save
button, it became necessary to allow the ORA to retry saving. This logic
mostly is around debouncing error banners so the user can continue to
edit a response without unnecessary defocusing or visual noise.