fix: Update frontend-build, webpack and karma-webpack.
Update to the latest frontend-build to pull in some webpack config
updates (https://github.com/openedx/frontend-build/pull/475). This
necessitated an update to Webpack which was out of date, which in-turn
resulted in an update to `karma-webpack`.  I pulled in the latest
changes and dealt with the one breaking change to karma config.