refactor: update assessments data format
Frontend now uses a data format which differs from how we store
assessments in the backend. Had to create a bunch of extra serializers
to handle these different views and directions of data transformation.

Ideally there is follow-up work here to minimize the number of one-off
serializers and consolidate some of that logic.