fix: Update webpack configs to transpile openedx packages.
This babel-loader config is pulled from frontend-build@13.0.4

We had to do this because frontend-build is currently too far behind and
can't easily be updated to a newer version without upgrading webpack and
many other libraries.

By making this small change we unblock the paragon upgrade to the new
scope and latest version.