chore: Update to the new version of paragon in the new scope. (#2099)
* chore: Update to the new version of paragon in the new scope.

Part of https://github.com/openedx/axim-engineering/issues/23

This replaces the `@edx/paragon` packag to point to the `paragon` package at
the `openedx` scope(`@openedx/paragon`). Imports have been updated to use the
same locations in the new package.

* fix: Update webpack configs to transpile openedx packages.

This babel-loader config is pulled from frontend-build@13.0.4

We had to do this because frontend-build is currently too far behind and
can't easily be updated to a newer version without upgrading webpack and
many other libraries.

By making this small change we unblock the paragon upgrade to the new
scope and latest version.

* chore: Static asset compilation.

* build: Add an nvmrc file to be clear about the node version.

This is now best practice in repos with JS code so that it's easy to be
using the right version of node.  This repo is still on node 16 because
edx-platform is still on node 16