fix: Disable file buttons when actions in progress
In the student view "Your Response" section,
there were several race conditions and issues relating to file uploads:

- Selecting the file input while multiple file uploads are in progress
  results in broken uploaded files and inconsistent UI.
- A file can be deleted after clicking Submit,
  while the submission is still processing.
- Clicking "Upload files" again after successfully uploading files,
  but before selecting more files,
  results in an attempt to upload the same files again.

On a high speed network connection,
the race condition issues are not very noticeable,
but on a slower connection,
they are easy to trigger as there is more time
in the UI while processing is happening in the backend.

These changes fix the above issues by ensuring that the buttons
and other input fields relating to file uploads
are disabled while actions are in progress
(uploading, deleting, submitting).

Private-ref: https://tasks.opencraft.com/browse/BB-9938